Register workers through the control socket
Once we add reforking, we won't be able to create a pipe
in the master before fork and have the worker inherit it.

So instead we create the pipe in the worker, and send it to
the master through the control socket.

NB: I'm not very happy with the `monitor_loop` extraction,
but this was necessary because the test suite rely on `start`
spawning the initial workers synchronously.