Start porting integration tests to Ruby
This commit introduces some of the foundation required to port the
entire integration suite to Ruby. These tests aren't yet run as part of
the suite, the entire bash suite can be removed when full parity is
achieved. It mostly mimicks the flow of the bash tests, but most of the
assertions are execute in the Ruby context (as opposed to using shell
constructs like cURL, grep etc.)