Allow to mark workers as fork unsafe
When some part of an application is hard or impossible to make
fork-safe, but rarely called, it's possible to simply mark the worker
as fork unsafe, and rely on its siblings to fork the next generation.

In the eventually where all the workers end up marked fork unsafe,
no new generation will be spawned, unless of course one of the unsafe
workers die and is replaced by a fresh one.

So this is only meant as a workaround for infrequently called things.