Add devcontainer configuration
This project already have a Dockerfile to make it easier to develop
since it depends on a lot of feature from linux. However, if users
develop using VSCode it would be good to have a devcontainer
config to make this Dockerfile just work.

This require changing the base image to the devcontainer images that
include the vscode server and other developer tools.