Pitchfork.fork_sibling: handle the middle process getting stuck
That middle process only forks again and then exit, as such it should
be very fast.

However because of after fork callbacks (on `Process._fork`) and
`at_exit` callbacks, it's possible that a bug cause it to get stuck
or be very slow.

To be resilient to this kind of bug, we hard kill the middle process
if it didn't exit in 2 seconds.

And to make it easier to debug, we kill it with `SIGBUS` so it produce
a crash report with a backtrace, core dump etc.