Handle fork_sibling failing repeatedly
Ref: https://github.com/Shopify/pitchfork/issues/79

It can happen that the new mold was forked while at an unsafe
point, causing the middle process to crash.

When we detect this happens, we should abandon this mold.

Currently abandoning the mold cause a graceful shutdown,
in the future we could try creating a new mold to replace it.

Co-Authored-By: Étienne Barrié <etienne.barrie@gmail.com>