Abort if worker consistently can't be spawned
Users can put anything in `after_worker_fork` and can cause
worker spawn to never succeed. Either because it's outright
broken or because it became slow enough to timeout often.

In such case it's probably best to just abort, otherwise
we may end up in an infinite loop of trying to spawn workers.