Avoid double response on post-response exception
Prior to this commit it was possible to send a double response if
anything raised after `http_response_write` but before the `rescue and
`handle_error` below it: https://github.com/Shopify/pitchfork/blob/18869d2f02549a54d7b2db6e0351e7fa71e95546/lib/pitchfork/http_server.rb#L878-L892

I'm not sure this is stricly a problem—it seems like most clients will
ignore the additional response? But it also seems unhelpful and
potentially confusing.

I don't think there are many ways for this to happen, but we did see it
in our application with `Rack::BodyProxy`, which allows calling a block
when the body is closed.

The test added in this commit previously failed because it wrote both a
success and an error response:

```
Expected

  /HTTP\/1.[01] 500\b/

  to not match

  "HTTP/1.1 200 OK\r\n
  Date: Thu, 30 Oct 2025 15:05:55 GMT\r\n
  Connection: close\r\n
  content-type: text/plain\r\n
  \r\n
  hello!\\n
  HTTP/1.1 500 Internal Server Error\r\n
  \r\n"
```