Fix after_request_complete getting nil env
Previously, when a request caused `@request.read(client)` to raise,
`after_request_complete` callbacks would end up with a `nil` `env`
(which can cause `NoMethodError`s when assuming its always a `Hash`).

This commit fixes the `NoMethodError` possibility by ensuring the `env`
returned is at least a `Hash` so that `[]` won't raise. Note that it
still can't be treated as a _valid_ Rack environment since the request
wasn't able to finish parsing, but a `Hash` here is better than `nil`.