Fix Rack 3.2 compatibility
One of the changes in Rack 3.2 was a fix to ensure at least one of
PATH_INFO or SCRIPT_NAME are non-empty. This causes a Rack::Lint error
in Pitchfork for `OPTIONS *` requests because Pitchfork has a `nil`
`PATH_INFO` for these.

This commit fixes the issue by setting `PATH_INFO` to `*`. It also undoes
some of the previous changes from the [commit][1] to support Rack 3.1
because they appear to have no effect.

The undone change is the moving of `PATH_INFO = "*"` and `REQUEST_PATH =
"*"` to the `fragment` action. From what I can tell, this part of the
state machine is not involved with the request uri, and so this if
statement is not evaluated in `OPTIONS *` requests:

```
  Request_Line = ( Method " " Request_URI ("#" Fragment){0,1} " " HTTP_Version CRLF ) ;
```

Then, the actual fix is manually setting the `PATH_INFO` to `*` when
that's the given `REQUEST_URI`. Normally, `PATH_INFO` is handled in the
`request_uri` action, but that action is only run for URIs starting with
`/` (and explicitly excludes `*`).

[1]: 65725c52b8170cbafa48551e952ce11ec0104bef